Ensure additional commits after approval are rejected (#11223)
When running this workflow manually (signalled via the `workflow_dispatch` event),
we need to ensure that we are operating on commits that have been
reviewed/approved. This is a safety measure to prevent:

1. Accidental pushes that might cause bugs.
2. Malicious pushes that might do nefarious things.

I considered implementing this by re-checking for approval at the time
we checkout the code, but because the `gh` call to check for approval is
a separate call from the checkout, we will always need to do this SHA
comparison to avoid a race. So the simplest thing was to memoize the SHA
as part of the output from the first job and avoid another `gh` call.

Lastly, this validation relies on `Dismiss stale pull request approvals
when new commits are pushed` being set in the repo settings. It's already
configured that way here in `dependabot-core`.