Bump Hex Erlang to 25
Currently we are using 24 from apt, and there are notes that we should not use ESL for this version of Ubuntu, but times have changed and ESL seems to have caught up.

I think ESL is the best way forward outside of using the RabbitMQ erlang solutions (which I'm fine with doing if desired, I'm just more familiar with ESL).

This also bumps Elixir by a patch version.