Update `DEPENDENCIES_LABEL_REGEX`
This regex is used to find and apply dependency labels to Dependabot pull requests. This feature is undocumented and sometimes results in unexpected and unwanted labels being applied to Dependabot pull requests.

Updating the regex to only match `depenency` or `dependencies` instead of the substring `dependenc` should fix this in most cases.