Match input EOLs when performing nuget updates
Currently, when using nuget to perform the updates the end-of-line
characters get reset to uniformly be \n. This is typically fine, except
for cases where project files are set to have \r\n line endings in the
.gitattributes file of a repo, where this results in dependabot making a
PR that breaks EOL normalization. To avoid this, we can simply cache the
original line endings and re-apply those after performing the update.