Make `setup-go` inherit `go` version from the `go.mod` file
This way it's one less place we have to keep things in-sync.

I realize the version in `go.mod` is intended as a minimum bound and not a hardcoded version, for example, the underlying file is already using a slightly different version:
https://github.com/dependabot/dependabot-core/blob/d0b3517a63291220cbdb7b04f1f6859474fcd375/silent/tests/go.mod

But practically speaking my experience is this way is simpler to reason about.