Document why `bundler-cache: true` isn't needed (#11711)
I was about to add `bundler-cache: true` when I realized that we don't actually `bundle install` anything. So the cache is not only superfluous, it's probably slower.

Since I spent a little bit of time digging into this and it's easy to miss if someone is in a hurry, I thought best to document for the next person.

Hopefully this way if someone changes to `bundle install` a gem, they'll realize this should get flipped on.