Publish gems using the version of RubyGems specified in `Dockerfile.updater-core` (#12036)
Rather than inheriting the version of RubyGems bundled with Ruby, let's
explicitly match the version of RubyGems specified in
`Dockerfile.update-core`, as this is what the rest of our prod/CI
systems are using. So this feels a bit safer.

It's also more flexible because it lets us explicitly manage the version
of RubyGems.

I think this is a good idea overall, but the specific need for this
arose because the `--attestations` flag was only recently added to the
`gem push` command. So we need a newer version of RubyGems than is
bundled with the version of Ruby currently used by this repo (3.3). It
turns out that we already use a newer version, so we just need to wire
that up into this workflow as well.