Only sleep on errors
Gem publishing was taking a surprisingly long time.

Turns out this is because the rakefile has a 2-second sleep between every
publish. I suspect this was originally implemented due to rate limits or
network flakiness, years ago when we only had a couple of gems.

However, RubyGems now implements separate rate limits for Gem push
calls, and they're pretty generous: https://guides.rubygems.org/rubygems-org-rate-limits/

So let's only sleep on errors.

The better fix is to switch to only looping/sleeping on transient errors like 429's.
But this isn't critical code for us and we rarely hit errors, so the
current retry implementation is good enough.