Add explicit dependency on `ostruct`
As of Ruby 3.4 `ostruct` has been deprecated and is no longer loaded by default. In Ruby 3.5 it will be completely removed.

As a stopgap solution before migration to Ruby classes or Hashes, we've added a direct dependency on the `ostruct` gem.

See #11163 for more information. This is part of the larger Ruby 3.4 migration in #11681