fix(common): return changelog url only if text exists
Currently, the changelog_url function (message_builder.rb) returns a
url to the first changelog file with a name from the list defined in
Dependabot::MetadataFinders::Base::ChangelogFinder::CHANGELOG_NAMES.

This causes it to sometimes return a url to a file, which is definitely
not a changelog (e.g. some script used to make a release of a library).
The commit message ends up containing a url that shouldn't be there.

The solution is to check the contents of the changelog file by invoking
the changelog_text function (metadata_finders/base.rb) in an 'if'
statement before returning the result of the changelog_url function.

The same approach is used in the changelog_cascade function, which is
used for building a pull request message (metadata_presenter.rb).

Signed-off-by: Oleksandr Zakharov <x.zakharov@proton.me>