Don't have a fallback on `"master"` as default
Some repos have `"main"`, some have `"trunk"`, and so it's not safe for us to default to `"master"`... 

Because if they have a different mainline branch, and _also_ a `"master"` branch, then we will pick the wrong one.

We _used_ to follow this `"master"` branch fallback pattern, but that was back before `"main"` became popular. So I've removed it from all the other ecosystems and things are working fine.

So removing it here as well.