`Dependabot::Dependency#top_level?` can be set by ivar
In order to create DependencySubmission payloads, we must keep track of
dependencies on a per-DependencyFile level, and those dependency objects
must know whether they are direct dependencies (`#top_level?`) and
whether they are usined in `production?`.

First we tried faking `Dependency#requirements` objects but this broke
a number of tests; somewhere deep within Dependabot there is an
assumption that lockfiles do not produce dependencies with requirements.

Instead of trying to fix the tests, here is a different approach:
- Instead of faking `requirements` objects, we instead directly supply
  the `top_level?` boolean at instantiation, i.e. before, top_level was
  a synonym for `requirements.any?` and now it is not.

Hopefully this turns out to be a simpler modification going forward.