Simplify how we attribute the global dependency list to a single file
We originally started building out per-file lists on the basis that returning
a holistic snapshot where we attribute direct dependencies to the manifest
and direct and indirect dependencies to the lockfile.

Since we only actually want to submit the most specific file to mirror existing
Dependency Graph behaviour, this ends up being a convolution we can remove -
especially now we have a better understanding of multi-directory support.

For a single directory, Dependabot builds the dependency set in a left-to-right
precedence from manifest to lockfile for most ecosystems, so the correct file
to attribute the dependency set to is always the right-most file in the set.

There may be some ecosystems where this assumption of deterministic ordering
falls down, but we can handle that by making the grapher a common/ component
and affording ecosystem overrides in the situations where this matters.