Allow for missing go directive in `go.mod`
We're seeing the following issue in some repositories

```
Passed `nil` into T.must
file_parser.rb:70:in 'Dependabot::GoModules::FileParser#go_version
```

The issue is that the `go` directive in `go.mod` files was optional before Go 1.21[1]. So some projects which rely on vendor dependencies that have not been updated since then will not have the `go` directive.

> The go directive sets the minimum version of Go required to use this module. Before Go 1.21, the directive was advisory only; now it is a mandatory requirement: Go toolchains refuse to use modules declaring newer Go versions.

This change makes it optional in Dependabot as well.

[1]: https://go.dev/doc/modules/gomod-ref#go