Prefer scm url for source url in maven (#2084)
Swap the order of project url and scm url as potential source urls to
ensure that scm url is preferred. Keep issue management url as the
lowest priority potential source url. Add tests verifying that the
correct source url is determined based on the presence or absence of scm
url, project url, and issue management url.