Handle situations where the OpenTofu registry errors
There are situations where the OpenTofu registry returns a non-200
status code for a reason (Such as a cloudflare outage!) And this commit
aims to improve the error logging around this to not confuse end users
of Dependabot with OpenTofu.

It adds in different status code handling branches to ensure that
the correct type of error and message is surfaced up to the end-user.

Signed-off-by: James Humphries <james@james-humphries.co.uk>