Streamline 34 `COPY` commands into one
Docker layers are useful for caching, but having too many layers can lead to inefficiencies in the build process and larger image sizes.

Additionally, within GitHub, when we pull images from the GitHub Container Registry,
each layer generates a separate download request, so we've been spamming that service will millions of calls.

Additionally, most of the time if one of these files changes, we end up invalidating the cache for all subsequent layers anyway.
Plus the time to `COPY` is very quick so there's little benefit to having them as separate layers.

All that to say, it makes sense to streamline these 34 `COPY` commands into a single command.

Previously this wasn't possible because Docker didn't offer a way for a single `COPY` command
to preserve nested file paths for individual files.

With the introduction of the new `--parents` flag, plus wildcards, we can achieve this.