Use a single `COPY` command within the python version builder images
Previously this was using multiple `COPY` commands.

Merging them should result in slightly smaller and faster builds, at
the theoretical tradeoff of busting the cache more frequently. But reality
is these layers will all either change at the same time, or not change, so
one layer vs three is identical for cache hits.

Each layer is also a separate call to the underlying docker registry so
by combining we are less likely to spam them and potentially trigger rate limits.
Parallelization is great for larger file sizes, but here the directory contents
are relatively small so the speed decrease should be minimal.