Handle multipe `COPY` into `$DEPENDABOT_HOME/` as one command (#13638)
This streamlines the `COPY` into `$DEPENDABOT_HOME/` as a single command.

This reduces the layer count from two (or more) to only one.

This shouldn't have _that_ much impact on the cache hits since the two layers were
adjacent. But it will reduce the calls to the docker registry since each layer is a separate call.
And should make for slightly faster / trimmer builds when building from scratch.