Use `NODE_EXTRA_CA_CERTS` for all Node package managers
The `NODE_EXTRA_CA_CERTS` env var was previously set with a comment that
it only applied to yarn berry. However, it actually applies to all Node.js
package managers, so that will include the yarn v1 package manager.

Setting this env var means we can remove the `.yarnrc` file that was
previously added as a workaround for yarn v1.

I also moved the `NODE_EXTRA_CA_CERTS` env var to earlier in the Dockerfiles
so that it's set before any package manager commands are run as part of the
native helpers build step.

The dockerfiles still mention hacky workarounds for npm 6, which I do
plan to address, but that is out of scope for this change.