Support dist for dev prefixed local packages
Institutes a fallback to reference the dist commit hash if source is missing. This will happen in path-based repositories, particularly. And because we don't conditionalize dev prefixed constraints, this will result in a mismatch of the constraint to the lockfile (dev-main# to dev-main).

The primary branch name is often chosen as the constraint in order to retain stability in the lockfile (it'll guarantee the lockfile's `version` is set to the main branch name rather than dynamically to the source branch of the `composer require` operation).